Reactor Essentials - External Matte Saver.fuse
The "External Matte Saver" fuse has been updated to work with Resolve 15. This provides a way to export multiple alpha channels from the Fusion page. If you like the idea of the External Matte Saver fuse and want more control on the channels that are exported then you should check out the "LifeSaver" fuse in the Reactor "Tools > IO" category.