KartaVR Scripts | Geometry | Send Geometry to usdview
The "Send Geometry to usdview" script takes AlembicMesh3D nodes that are selected in the Nodes view and sends the .abc models to usdview. The usdview program needs to be installed on the system and listed in the $PATH environment variable. usdview is primarily used to examine PIXAR OpenUSD (USDC/USDA) format scene graphs but it can also work with Alembic models too.