KartaVR Scripts | Virtual Production
The KartaVR "Virtual Production" atom allow you optimize you on-set VFX workflows and interact with match moving department provided information like camera data, point clouds, and stand-in models. This is essential for effective scene integration and virtual produciton efforts in Fusion. The scripts in this atom package are the rare and exotic "rosetta stone" like tools that will allow your pipeline to interchange Fusion matchmoving data with external DCC apps, and make it a painless process to export PIXAR USD ASCII and Maya ASCII formatted copies of your Fusion 3D scene graph based nodes. This atom package is a beta toolset that will evolve and change in the leadup to KartaVR v5 in 2020.